Unable to find license details
Could not find license details in the readme or a separate license file. Not having a license hinders adoption.

I have updated the readme file to mention MIT license, if it is acceptable to you.

Mentioning the license at end of the readme file seems to be becoming the norm.

Thanks.